Add files via upload
Clean code - version stable
